Removed few warnings in code

+++ VKRenderer/Applications/ComputeCull/ComputeCull.cpp
@@ -620,7 +620,7 @@ void ComputeCull::PrepareApp()
 
 }
 
-void ComputeCull::Update(float deltaTime)
+void ComputeCull::Update(double deltaTime)
 {
 	m_renderer->ProcessInput(m_renderer->m_window, deltaTime);
 
@@ -632,7 +632,7 @@ void ComputeCull::Update(float deltaTime)
 	return;
 }
 
-void ComputeCull::Draw(float deltaTime)
+void ComputeCull::Draw(double deltaTime)
 {
 	vkWaitForFences(m_renderer->m_device, 1, &m_renderer->m_inflightFences[m_currentFrame], VK_TRUE, std::numeric_limits<uint64_t>::max());
 	vkResetFences(m_renderer->m_device, 1, &m_renderer->m_inflightFences[m_currentFrame]);
